initialize views after onCreate, change MovieAdapter and it's onClickListener todo: Caused by: java.lang.NullPointerException: Attempt to invoke virtual method 'com.annamorgiel.popularmovies_udacity_1.Rest.service.MovieService com.annamorgiel.popularmovies_udacity_1.Rest.RestClient.getMovieService()' on a null object reference                                                                                              at com.annamorgiel.popularmovies_udacity_1.ui.DetailActivity.fetchMovieDetails(DetailActivity.java:59)                                                                                              at com.annamorgiel.popularmovies_udacity_1.ui.DetailActivity.onCreate(DetailActivity.java:55)

+++ app/src/main/java/com/annamorgiel/popularmovies_udacity_1/Rest/service/MovieService.java
@@ -26,6 +26,6 @@ public interface MovieService {
     Call<ApiResponse> getMovies(@Path("sortby") String sortby, @Query("api_key") String apiKey);
 
     @GET("{id}")
-    Call<ApiResponse> getMovieDetails(@Path("id") Long id, @Query("api_key") String name);
+    Call<ApiResponse> getMovieDetails(@Path("id") Integer id, @Query("api_key") String name);
 
 }
